Refactoring continued.
