+++ src/sys/mod.rs
@@ -1,0 +1,3 @@
+pub mod memory;
+pub mod processor;
+pub mod storage;
